+++ src/modules/motion.cpp
@@ -91,7 +91,7 @@ void Motion::StallGuardReset(Axis axis) {
     axisData[axis].drv.ClearStallguard();
 }
 
-void Motion::PlanStallGuardThreshold(config::Axis axis, int16_t sg_thrs) {
+void Motion::PlanStallGuardThreshold(config::Axis axis, int8_t sg_thrs) {
     mm::axisParams[axis].params.sg_thrs = sg_thrs;
 }
